# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  request.ReviewWhitelistRequest:
    properties:
      admin_notes:
        type: string
      approved:
        type: boolean
    required:
    - approved
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@eventcampus.com
    name: API Support
  description: |-
    REST API for Event Campus - Campus Event Management System

    Features:
    - User authentication (Register, Login)
    - Event management (Create, Read, Update, Delete)
    - Event registration
    - Attendance tracking
    - Whitelist/Organizer approval system
    - Email notifications
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Event Campus API
  version: "1.0"
paths:
...
  /events/{id}/register:
    post:
      consumes:
      - application/json
      description: Register authenticated user for a specific event
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Registration successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid event ID or registration failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Register for an event
      tags:
      - Registrations
  /events/{id}/registrations:
    get:
      consumes:
      - application/json
      description: Get all registrations for a specific event (organizer only)
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Registrations retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid event ID or failed to get registrations
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get event registrations
      tags:
      - Registrations
...
  /registrations/{id}:
    delete:
      consumes:
      - application/json
      description: Cancel user's registration for an event
      parameters:
      - description: Registration ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Registration cancelled successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid ID or cancellation failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel event registration
      tags:
      - Registrations
  /registrations/my:
    get:
      consumes:
      - application/json
      description: Get all registrations for authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Registrations retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get registrations
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my registrations
      tags:
      - Registrations
  /whitelist/{id}/review:
    patch:
      consumes:
      - application/json
      description: Approve or reject whitelist request (admin only)
      parameters:
      - description: Request ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Review decision
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ReviewWhitelistRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Request reviewed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or review failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Review whitelist request
      tags:
      - Whitelist
  /whitelist/my-request:
    get:
      consumes:
      - application/json
      description: Get authenticated user's whitelist request status
      produces:
      - application/json
      responses:
        "200":
          description: Request retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "404":
          description: No request found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my whitelist request
      tags:
      - Whitelist
  /whitelist/request:
    post:
      consumes:
      - multipart/form-data
      description: Submit request to become event organizer (requires PDF document)
      parameters:
      - description: Organization name
        in: formData
        name: organization_name
        required: true
        type: string
      - description: Supporting document (PDF only)
        in: formData
        name: document
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Request submitted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or upload failed
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Submit whitelist request
      tags:
      - Whitelist
  /whitelist/requests:
    get:
      consumes:
      - application/json
      description: Get all whitelist requests with optional status filter (admin only)
      parameters:
      - description: Filter by status (pending/approved/rejected)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Requests retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get requests
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all whitelist requests
      tags:
      - Whitelist
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"